\name{clustering}
\alias{clustering}

\title{Computes the similarity matrix}

\description{
Computes the similarity matrix.
}

\usage{clustering(object, ...)}

\arguments{
\item{object}{an object of class "mvrm", usually a result of a call to \code{mvrm}.}
\item{...}{other arguments.}
}

\details{The function computes the similarity matrix for clustering based on corrrelations or variables.}

\value{Similarity matrix.}

\author{Georgios Papageorgiou \email{gpapageo@gmail.com}}

\examples{#see \code{mvrm} example}

\seealso{\code{\link{mvrm}}}
