\name{histCorr}
\alias{histCorr}

\title{Creates plots of the correlation matrices}

\description{
This function plots the posterior distribution of the elements of correlation matrices. 
}

\usage{histCorr(x, term="R", plotOptions = list(),...)}

\arguments{
\item{x}{an object of class `mmvrm', as generated by function \code{mmvrm}.}
\item{term}{Admits two possible values: "R" to plot samples from the posterior of \eqn{R} and "muR" to plot samples from the posterior of \eqn{\mu_R}.}
\item{plotOptions}{ggplot type options.}
\item{...}{other arguments.}
}

\details{Use this function to visualize the elements of a correlation matrix.}

\value{Posterior distributions of elements of correlation matrices.}

\author{Georgios Papageorgiou \email{gpapageo@gmail.com}}

\examples{#see \code{mvrm} example}

\seealso{\code{\link{mvrm}}}
