\name{plotCorr}
\alias{plotCorr}

\title{Creates plots of the correlation matrices}

\description{
This function plots the posterior mean and credible intervals of the elements of correlation matrices. 
}

\usage{plotCorr(x, term="R",centre = mean, quantiles = c(0.1, 0.9), ...)}

\arguments{
\item{x}{an object of class `mmvrm' as generated by function \code{mmvrm}.}
\item{term}{R or muR,}
\item{centre}{a description of how the centre of the posterior should be measured. Usually \code{mean} or \code{median}.}
\item{quantiles}{the quantiles to be used when plotting credible regions. 
Plots without credible intervals may be obtained by setting this argument to NULL.}
\item{...}{other arguments.}
}

\details{Use this function to visualize the elements of a correlation matrix.}

\value{Posterior means and credible intervals of elements of correlation matrices.}

\author{Georgios Papageorgiou \email{gpapageo@gmail.com}}

\examples{#see \code{mvrm} example}

\seealso{\code{\link{mvrm}}}
