\name{simD}

\alias{simD}

\docType{data}

\title{Simulated dataset}

\description{Just a simulated dataset to illustrate the DO mixture model. The success probability and the covariate have a non-linear relationship.}

\usage{data(simD)}

\format{
  A data frame with 300 independent observations. Three numerical vectors contain information on
  \describe{
    \item{\code{Y}}{number of successes.}
    \item{\code{E}}{number of trials.}
    \item{\code{X}}{explanatory variable.}
  }
}


\keyword{datasets}
