\name{simD2}

\alias{simD2}

\docType{data}

\title{Simulated dataset}

\description{A simulated dataset to illustrate the multivariate longitudinal model. It consists of a bivariate vector of responses observed over 6 time points.}

\usage{data(simD2)}

\format{
  A data frame that includes observations on 50 sampling units. The data frame has 300 rows for the 50 sampling units observed over 6 time points. It has 4 columns 
  \describe{
    \item{\code{Y1}}{first response.}
    \item{\code{Y2}}{second response.}
    \item{\code{time}}{the time of observation.}
    \item{\code{id}}{unique sampling unit identifier.}
  }
}

\keyword{datasets}
