\name{BNSP-package}

\alias{BNSP-package}
\alias{BNSP}

\docType{package}

\title{Bayesian non- and semi-parametric model fitting}

\description{Markov chain Monte Carlo algorithms for non- and semi-parametric models: 1. Dirichlet process mixture models with function \code{dpmj} and 2. spike-slab variable selection in mean/variance regression models with function \code{mvrm}.}

\details{
\tabular{ll}{
Package: \tab BNSP\cr
Type: \tab Package\cr
Version: \tab 2.0.7\cr
Date: \tab 2018-09-27\cr
License: \tab GPL (>=2)\cr}

This program is free software; you can redistribute it and/or
modify it under the terms of the GNU General Public License
as published by the Free Software Foundation; either version 2
of the License, or (at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

For details on  the GNU General Public License
see \url{http://www.gnu.org/copyleft/gpl.html} or
write to the Free Software Foundation, Inc., 51 Franklin Street,
Fifth Floor, Boston, MA  02110-1301, USA.}

\author{
Georgios Papageorgiou (2014)

Maintainer: Georgios Papageorgiou <gpapageo@gmail.com>}

\section{Acknowledgments}{This work was partly supported by the Medical Research Council grant number G09018401.}

\references{
Papageorgiou, G. (2018). Bayesian density regression for discrete outcomes. arXiv:1603.09706v3 [stat.ME]. 

Papageorgiou, G. (2018). BNSP: an R Package for Fitting Bayesian Semiparametric Regression Models and Variable Selection. arXiv:1804.10939 [stat.OT]

Papageorgiou, G., Richardson, S. and Best, N. (2015). Bayesian nonparametric models for spatially indexed data of mixed type.
Journal of the Royal Statistical Society: Series B (Statistical Methodology), 77:973-999.}
