\name{BNSP-package}

\alias{BNSP-package}
\alias{BNSP}

\docType{package}

\title{Bayesian non- and semi-parametric model fitting}

\description{Markov chain Monte Carlo algorithms for non- and semi-parametric models}

\details{
\tabular{ll}{
Package: \tab BNSP\cr
Type: \tab Package\cr
Version: \tab 1.0.5\cr
Date: \tab 2015-07-08\cr
License: \tab GPL (>=2)\cr}

This program is free software; you can redistribute it and/or
modify it under the terms of the GNU General Public License
as published by the Free Software Foundation; either version 2
of the License, or (at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

For details on  the GNU General Public License
see \url{http://www.gnu.org/copyleft/gpl.html} or
write to the Free Software Foundation, Inc., 51 Franklin Street,
Fifth Floor, Boston, MA  02110-1301, USA.}

\author{
Georgios Papageorgiou (2014)

Maintainer: Georgios Papageorgiou <gpapageo@gmail.com>}

\section{Acknowledgments}{This work was partly supported by the Medical Research Council grant number G09018401.}

\references{
Papageorgiou, G., Richardson, S. and Best, N. (2014). Bayesian nonparametric models for spatially indexed data of mixed type.
Accepted: Journal of the Royal Statistical Society - Series B.}
